+++ src/components/AllBooks.js
@@ -1,0 +1,5 @@
+export default function AllBooks() {
+  return `<li>
+              <button class="${styles.btn}" type='button' onclick="showAllBooks()">Show all</button>
+          </li>`;
+}
