+++ src/components/Randomiser.js
@@ -1,0 +1,3 @@
+export default function Randomiser() {
+  return `<p>Wondering what to read? <button class="${styles.btn}" type='button' onclick="getRandomBooks(randomWord())">Get advise</button></p>`;
+}
