+++ src/data/dataStore.js
@@ -1,0 +1,12 @@
+window.dataStore = {
+  isNotFiltered: true,
+  isDataLoading: false,
+  theBooksInfo: [],
+  filteredBooks: [],
+  books: [],
+  random: '',
+  filterValue: '',
+  searchValue: '',
+};
+
+export default dataStore;
