+++ src/framework/render.js
@@ -1,0 +1,5 @@
+import App from '../components/App';
+
+export default function renderApp() {
+  document.getElementById('app-root').innerHTML = ` ${App()} `;
+}
